+++ java/lilithscythemod/Skill/SkillNBTData.java
@@ -15,24 +15,24 @@ public class SkillNBTData implements IExtendedEntityProperties{
 	public static final String Tag_CoolTime	= "CoolTime";
 	public static final String Tag_GlobalCoolDown ="GlobalCoolDown";
 	public static final String Tag_ChargeTime ="ChargeTime";
-	
+
 	private byte stateByte = 0x01;
 	private NBTTagList skillList = new NBTTagList();
-	
+
 	/**スキルの詠唱状態
 	 * Ready:スキル発動可能状態
-	 * 
+	 *
 	 * Chant:スキル詠唱中
 	 * ChantEnd:スキル詠唱完了
-	 * 
+	 *
 	 * GCD:グローバルクールダウン中
 	 * CoolDown:クールダウン中*/
 	public static final byte State_Ready		= 0x01;
 	public static final byte State_Chant		= 0x02;
 	public static final byte State_ChantEnd		= 0x03;
 	public static final byte State_CoolDown		= 0x04;
-	
-	
+
+
 	@Override
 	public void saveNBTData(NBTTagCompound compound) {
 		 compound.setTag(NBTNAME, skillList);
@@ -42,26 +42,26 @@ public class SkillNBTData implements IExtendedEntityProperties{
 	public void loadNBTData(NBTTagCompound compound) {
 		this.skillList = compound.getTagList(NBTNAME, Constants.NBT.TAG_COMPOUND);
 	}
-	protected void addSkillData(String skillName,int coolTime,float ChargeTime,byte stateByte){
+	protected void addSkillData(String skillName,float coolTime,float ChargeTime,byte stateByte){
 		 NBTTagCompound nbt = new NBTTagCompound();
 		 nbt.setByte(Tag_State, stateByte);
 		 nbt.setString(Tag_SkillName, skillName);
-		 nbt.setInteger(Tag_CoolTime, coolTime);
+		 nbt.setFloat(Tag_CoolTime, coolTime);
 		 nbt.setFloat(Tag_ChargeTime, ChargeTime);
 		 skillList.appendTag(nbt);
 	}
-	
+
 	@Override
 	public void init(Entity entity, World world) {}
-	
+
     public static void register(EntityLivingBase entity) {
     	entity.registerExtendedProperties(NBTNAME, new SkillNBTData());
     }
-    
+
     public static SkillNBTData get(EntityLivingBase entity){
         return (SkillNBTData)entity.getExtendedProperties(NBTNAME);
     }
-    
+
 	public NBTTagList getNBTList()
 	{
 		return skillList;
@@ -86,7 +86,7 @@ public class SkillNBTData implements IExtendedEntityProperties{
 	{
 		return getNBTCompound(index).getFloat(key);
 	}
-	
+
 	protected void setNBTInt(int index,String key,int newValue)
 	{
 		getNBTCompound(index).setInteger(key, newValue);
@@ -96,7 +96,7 @@ public class SkillNBTData implements IExtendedEntityProperties{
 	{
 		return getNBTCompound(index).getInteger(key);
 	}
-	
+
 	protected void setNBTByte(int index,String key,byte newValue){
 		getNBTCompound(index).setByte(key, newValue);
 	}
